Move test helper modules into tests/helper path.
